added func to show every user time and roles by time

+++ src/bot.js
@@ -9,9 +9,6 @@ const client = new Client({
 const prefix = "$";
 
 client.on("voiceStateUpdate", (oldMember, newMember) => {
-  if (newMember.member.user.id !== oldMember.member.user.id) {
-    console.log("turess");
-  }
   const newUser = new UserTime(oldMember, newMember);
   newUser.time();
 });
@@ -21,55 +18,46 @@ client.on("ready", () => {
 });
 
 client.on("message", async (msg) => {
-  if (msg.author.bot) return;
-  if (msg.content.startsWith(prefix)) {
-    const [CMD_NAME, ...args] = msg.content
-      .trim()
-      .substring(prefix.length)
-      .split(/\s+/);
-    console.log(CMD_NAME, args);
+  try {
+    const jsonData = await utiles.readingFileSync("usersData.json");
 
-    // if (CMD_NAME === "role") {
-    //   msg.reply("choose your roles");
-    //   if (msg.content === "choose your roles") {
-    //     console.log(msg.reactions);
-    //   }
-    // }
-    if (CMD_NAME === "summon") {
-      msg.reply(`zamknij leb, już wbijam ${msg.author.username}`);
-      // msg.member.voice.channel.join();
-    }
+    if (msg.author.bot) return;
+    if (msg.content.startsWith(prefix)) {
+      const [CMD_NAME, ...args] = msg.content
+        .trim()
+        .substring(prefix.length)
+        .split(/\s+/);
+      console.log(CMD_NAME, args);
 
-    if (CMD_NAME === "time") {
-      try {
-        const jsonData = await utiles.readingFileSync("usersData.json");
+      // if (CMD_NAME === "role") {
+      //   msg.reply("choose your roles");
+      //   if (msg.content === "choose your roles") {
+      //     console.log(msg.reactions);
+      //   }
+      // }
+      if (CMD_NAME === "summon") {
+        msg.reply(`zamknij leb, już wbijam`);
+        // msg.member.voice.channel.join();
+      }
+
+      if (CMD_NAME === "time" && !args.length) {
+        console.log(!args.length);
+
         //check whitch user is doing this
 
         if (utiles.isEmpty(jsonData)) {
           msg.delete();
           msg.reply("plik jest pusty, upewnij się, że dołączyłeś na kanał");
         }
-        const arr = [];
-        for (let key of jsonData) {
-          let check =
-            msg.author.username === key.userData.name &&
-            msg.author.id === key.userData.useriD;
 
-          arr.push(await check);
-        }
+        const dataObj = await utiles.findUser(jsonData, msg);
 
-        let index;
-        const specCase = arr.filter((val) => {
-          const i = arr.findIndex((val) => val === true);
-          index = i;
-          return val;
-        });
+        const everyFalse = dataObj.arr.every(
+          (currentValue) => currentValue === false
+        );
 
-        const everyFalse = arr.every((currentValue) => currentValue === false);
-
-        if (specCase[0] === true) {
-          const diff = jsonData[index].userData.userTimeDiff;
-          console.log("diff", diff);
+        if (dataObj.specCase[0] === true) {
+          const diff = jsonData[dataObj.index].userData.userTimeDiff;
 
           const time = utiles.timeCounter(diff);
           utiles.format(time);
@@ -80,16 +68,55 @@ client.on("message", async (msg) => {
             msg,
             time.sec === 0 && time.min === 0
               ? "nie byłes jeszcze na żadnym kanale"
-              : utiles.format(time)
+              : utiles.format(time),
+            jsonData[dataObj.index].userData.name
           );
         } else if (everyFalse) {
           msg.delete();
           msg.reply("nie dołączyłeś jeszcze na żaden kanał");
         }
-      } catch (error) {
-        console.log(error);
+      }
+
+      if (CMD_NAME === "time" && args[0] === "all") {
+        const id =
+          msg.author.id === "278950668558139392" ||
+          msg.author.id === "194098078314266625";
+
+        jsonData.forEach((el) => {
+          if (id) {
+            const diff = el.userData.userTimeDiff;
+            const time = utiles.timeCounter(diff);
+            const format = utiles.format(time);
+            utiles.checkTime(
+              msg,
+              time.sec === 0 && time.min === 0
+                ? "nie byłes jeszcze na żadnym kanale"
+                : format,
+              el.userData.name ? el.userData.name : msg.author.username
+            );
+          }
+        });
+        if (!id) {
+          msg.reply("nie masz uprawnień");
+        }
+      }
+
+      if (CMD_NAME === "upgrade") {
+        const dataObj = await utiles.findUser(jsonData, msg);
+        if (dataObj.index === -1) {
+          msg.delete();
+          msg.reply("nie byłeś na żadnym kanale głosowym");
+        } else {
+          const diff = jsonData[dataObj.index].userData.userTimeDiff;
+          const id = jsonData[dataObj.index].userData.useriD;
+          msg.delete();
+
+          utiles.addRoleByTime(diff, id, msg);
+        }
       }
     }
+  } catch (error) {
+    console.log(error);
   }
 });
